+++ packages/eslint-config-vtex/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Explicitly set a `jest` version for rule `jest/no-deprecated-functions` to work.
 
 ## 12.8.9 - 2020-11-18
 ### Changed